Split loader into old K3D loader, and newly-implemented RTDL one
